+++ Engine/src/Graphics/SpriteComponent.h
@@ -36,7 +36,7 @@ namespace IonixEngine {
 
 	public:
 		SpriteComponent(Entity* entity, std::string alias, int width, int height, int zedOrder);
-		SpriteComponent(Entity* entity, uint64_t hash, int width, int height, int zedOrder);
+		SpriteComponent(Entity* entity, uint32_t hash, int width, int height, int zedOrder);
 		virtual void Render(RenderData* data) override;
 
 		void calculateTotalFrames();
